E3: Will try again tomorrow. Read descrip.
My source was kind of difficult for me to understand as I was retyping on atom and attempting to comprehend as I go. But an idea came to mind; maybe i can do something with image(CORNERS)? To make the background follow the user's movements.. Ok will attempt this tomorrow, I need to rest.