E3: Flipped the image of the both hibiscus and bee
I flipped the images and gave them a new x and y axis. The bee now begins from the right, and the hibiscus now sits in the bottom left corner. So now the bee will have to avoid the birds from eating it; the birds forging, as always, from left to right.