P2: Incredible! The stars are finally floating in a 3D map.
I followed Pippin's instructions about adding a z-axis in the Star.js' constructor, with negative and positive values in random(). I'm still not sure how p5 recognizes it as its "depth" but I assume that adding a 3rd axis automatically reads it as the 3rd dimension. Anyway, super exciting to see the stars move and take space (no pun intended) in the 3D simulation!!