P2: Set up camera movement in draw function so it followers user. Still needs to be improved though.
I created the variables “camX” and “camY” with a default value of 0, in script.js. We can find these variables in the parameters of p5’s 3D camera, under draw, and  in the Star.js class object under the “motion” method. The camera’s movement functions are similar to that of the user’s. That is,  “camX += this.speed” instead of “this.x += this.speed”. I figured that if I wanted the camera to move with the spaceship (user), why not program it similarly. It works and honestly I was super excited to see it go but there is space for improvement. Maybe I’ll define a separate handle input for the camera.