P2: Framed the scoring system of when a star is collected though there are issues that need to be improved.
So far, I've only established the 'skeleton' of how the scoring system will be programmed in the simulation. For testing purposes, the score increases when the “X” key is pressed. It works. Moving forward, I will have the scoring system correlate to when the user collects a star.

In an attempt to display a fixed element within the context of a camera that can be manipulated by the user, I used the function “createGraphics” in hopes of being able to set the element in a corner (and have it remain visible at all times) since using the rect( ) function didn’t work. I was fully aware that the camera controlled by the user was most likely (ok– 100%) the main issue but I still wanted to try in case this was the solution.