P2: Established Saturn's rings
Following Sam’s feedback from my progress report, I was able to display the rings of Saturn after previous failed attempts. Finally! I added the rings as an option in the superclass of all the planets. Inside the Planet.js parameter, includes a new parameter called “hasRings”. This is then customized in the script (in each subclass’ parameters) with boolean values of either true or false.