P2: User get to read about the planet if they were to approach it close enough
I really gave it a go on trying to have the information’s display dependent on the mousePress function but all my attempts failed. The closest I got was having all the spheres react at one click instead of individually. Even when reliant on mouseX and mouseY’s position in the sphere.

Finally, I decided to move forward with Pippin’s feedback and have the information appear when the user is close enough to a specific planet.